Revert "feat: reduced amount of re-renderings"
This reverts commit 4644f3ccad18f3713ae2831d7315455ba491dfd1.

+++ lib/src/timeline_painter.dart
@@ -99,7 +99,7 @@ class TimelinePainter extends CustomPainter {
   @override
   bool shouldRepaint(covariant CustomPainter oldDelegate) {
     if (oldDelegate is TimelinePainter) {
-      return value.toInt() != oldDelegate.value.toInt();
+      return value != oldDelegate.value;
     }
 
     return false;
